+++ str/ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: gpoblon <gpoblon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/09/17 18:14:49 by gpoblon           #+#    #+#             */
-/*   Updated: 2016/09/23 16:41:05 by gpoblon          ###   ########.fr       */
+/*   Updated: 2016/09/27 12:14:49 by gpoblon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strdup(const char *s)
 {
 	char *s_dup;
-	int i;
+	size_t i;
 
 	s_dup = ((char*)malloc(sizeof(char) * ft_strlen(*s) + 1);
 	if (*s_dup == NULL || *s == '\0')
